asmdef を整理する
