+++ Assets/ArtNet/Core/ArtNet.Core.asmdef
@@ -1,7 +1,7 @@
 {
-    "name": "ArtNetAssembly",
+    "name": "ArtNet.Core",
     "rootNamespace": "ArtNet",
-    "references": ["GUID:e88422c88230bca4d8e836cfcef4dd58"],
+    "references": [],
     "includePlatforms": [],
     "excludePlatforms": [],
     "allowUnsafeCode": false,
