+++ Assets/ArtNet/Samples/Scenes/SampleScene.unity
@@ -515,6 +515,7 @@ GameObject:
   m_Component:
   - component: {fileID: 1302849931}
   - component: {fileID: 1302849932}
+  - component: {fileID: 1302849933}
   m_Layer: 0
   m_Name: ArtNet
   m_TagString: Untagged
@@ -550,6 +551,19 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   bindIpAddress: 0.0.0.0
+--- !u!114 &1302849933
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1302849929}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: e3ecc7efcd1b4cf9ba8f5921cb165007, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  _artNetReceiver: {fileID: 1302849932}
 --- !u!1 &2128396738
 GameObject:
   m_ObjectHideFlags: 0
