+++ Assets/ArtNet/Scripts/ArtNet.Script.asmdef
@@ -1,7 +1,7 @@
 {
-    "name": "ArtNetCore",
+    "name": "ArtNet.Script",
     "rootNamespace": "ArtNet",
-    "references": [],
+    "references": ["GUID:e88422c88230bca4d8e836cfcef4dd58"],
     "includePlatforms": [],
     "excludePlatforms": [],
     "allowUnsafeCode": false,
